Added some color
You can now color selected vertices and edges with the click of a
button, or select all elements of a given color.  Six beautiful colors
to choose from.  Minor aesthetic improvements: improved the font, made
text un-selectable, changed the ‘pinned’ and ‘selected’ styles to be a
bit more lovely.  Set the stage for some buttons to modify the graph,
but have not implemented any such functionality.