Multi-edges and loops work with edge contraction
Also, loops are displayed (and cannot be curved), and the display of multi-edges is improved.  You can add loops from the interactive adjacency matrix and loops are ignored by the graph6 writer.  TikZ output and matrix input/output are still to come.  Link-backgrounds are a little smaller to make selecting multi-edges easier, link strength is only non-zero for one edge per multi-edge, so that multi-edge springiness is the same as single edge springiness.